Revert "feat: Introduce NodeBuilder (#5824)"
This reverts commit 8e9b02f128d7b485bd357b71fdc58a0edf54fbb0.
