+++ bin/reth/src/cli/mod.rs
@@ -21,9 +21,7 @@ use std::{fmt, fmt::Display, sync::Arc};
 
 pub mod components;
 pub mod config;
-pub mod db_type;
 pub mod ext;
-pub mod node_builder;
 
 /// Default [directives](Directive) for [EnvFilter] which disables high-frequency debug logs from
 /// `hyper` and `trust-dns`
